requirements.txt
Flask==0.10
Jinja2==2.6
gunicorn==19.1.1
Flask-Bootstrap==3.3.5.6
wtforms==2.0.2
Flask-WTF==0.11
requests==2.2.0
BeautifulSoup==3.2.1
geocoder==1.5.0
Flask-SQLAlchemy==2.0
psycopg2==2.5.3
Flask-Cors==2.1.0
user-agents==1.0.1
Flask-Login==0.3.2
pytz==2014.4
beautifulsoup4==4.3.2
Flask-Mail==0.9.1
python-dateutil==2.4.2
textblob==0.8.4
dnspython==1.12.0
